Moved wait kernel to common
